requirements.txt
streamlit
python-dotenv
google-genai
pydantic>=2.0
moviepy
pydub
fpdf2